Refactor ad description handling and update configuration structure
- Updated Configuration Structure:
  - Changed the ad description configuration from a nested format to a flattened format in `config_defaults.yaml`.

- Refactored Description Affix Logic:
  - Moved the `get_description_affixes` function from the `description` module to the `ads` module, consolidating related functionality.

- Code Cleanup:
  - Removed the now-unnecessary `description.py` module.
  - Updated imports across the codebase to reflect the new location of `get_description_affixes`.
  - Cleaned up the `KleinanzeigenBot` class to use the new function for retrieving description prefixes and suffixes.
